+++ samples/Demo/Beef.Demo.Test/FixtureSetup.cs
@@ -18,10 +18,9 @@ namespace Beef.Demo.Test
             TestSetUp.SetDefaultLocalReferenceData<IReferenceData, ReferenceDataAgentProvider, IReferenceDataAgent, ReferenceDataAgent>();
             TestSetUp.RegisterSetUp(async (count, data) =>
             {
-                return await DatabaseExecutor.RunAsync(
-                    count == 0 ? DatabaseExecutorCommand.ResetAndDatabase : DatabaseExecutorCommand.ResetAndData, 
-                    config["ConnectionStrings:BeefDemo"], useBeefDbo: true,
-                    typeof(Database.Program).Assembly, Assembly.GetExecutingAssembly(), typeof(Beef.Demo.Abc.Database.Scripts).Assembly).ConfigureAwait(false) == 0;
+                return await DatabaseExecutor.RunAsync(new DatabaseExecutorArgs(
+                    count == 0 ? DatabaseExecutorCommand.ResetAndDatabase : DatabaseExecutorCommand.ResetAndData, config["ConnectionStrings:BeefDemo"], 
+                    typeof(Database.Program).Assembly, Assembly.GetExecutingAssembly(), typeof(Beef.Demo.Abc.Database.Scripts).Assembly) { UseBeefDbo = true, RefDataSchemaName = "Ref" }).ConfigureAwait(false) == 0;
             });
         }
     }