+++ tools/Beef.Database.Core/CHANGELOG.md
@@ -3,7 +3,8 @@
 Represents the **NuGet** versions.
 
 ## v4.1.2
-- *Enhancement:* Latest `Beef.Core` and `Beef.CodeGen.Core` references needed.
+- *Enhancement:* The `ScriptNew` option has been changed to add support for a new `-createref` to create using the reference data schema; instead of the inferring from the schema name (this support has been removed). This way the developer explicitly decides and allows reference data to be used in any schema.
+- *Fixed:* Issue [71](https://github.com/Avanade/Beef/issues/71) has been resolved. A runtime error will now correctly result in a return code of `-1`.
 
 ## v4.1.1
 - *Enhancement:* Introduction of Dependency Injection support.